area_predictions_for_protected_landscapes/protected_landscapes_requirements.txt
access==1.1.8
affine==2.3.1
alembic==1.8.1
alphashape==1.3.1
ansi2html==1.8.0
appdirs==1.4.4
appnope==0.1.3
astor==0.8.1
asttokens==2.0.8
attrs==22.1.0
aws-requests-auth==0.4.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blosc2==2.2.9
bokeh==2.4.3
boto3==1.24.63
botocore==1.27.63
branca==0.5.0
Brotli==1.0.9
Cartopy==0.22.0
category-encoders==2.5.0
cattrs==22.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
clean-text==0.6.0
cleantext==1.1.4
click==8.1.3
click-log==0.4.0
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==2.1.0
colorcet==3.0.1
commonmark==0.9.1
ConfigArgParse==1.7
contextily==1.2.0
cramjam==2.6.2
cycler==0.11.0
cykhash==2.0.0
Cython==0.29.32
dash==2.6.1
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2023.6.0
databricks-cli==0.17.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
descartes==1.1.0
dill==0.3.5.1
distributed==2023.6.0
docker==6.0.0
docopt==0.6.2
emoji==1.7.0
entrypoints==0.4
esda==2.4.3
et-xmlfile==1.1.0
exceptiongroup==1.0.4
executing==1.0.0
factor-analyzer==0.4.0
fastjsonschema==2.16.1
fastparquet==2023.4.0
filelock==3.12.4
Fiona==1.8.21
Flask==2.2.2
Flask-Compress==1.12
Flask-Cors==3.0.10
folium==0.12.1.post1
fonttools==4.37.1
formulaic==0.3.4
fsspec==2023.5.0
ftfy==6.1.1
future==0.18.2
fuzzymatcher==0.0.6
fuzzywuzzy==0.18.0
gensim==4.2.0
geocoder==1.38.1
geodatasets==2023.3.0
geographiclib==1.52
geopandas==0.14.0
geoparquet==0.0.3
geoplot==0.5.1
geopy==2.2.0
gevent==21.12.0
giddy==2.3.3
gitdb==4.0.9
GitPython==3.1.27
graphviz==0.20.1
greenlet==1.1.3
gunicorn==20.1.0
h3==3.7.6
h3pandas==0.2.4
hdbscan==0.8.29
holoviews==1.15.1
hover==0.7.0
hvplot==0.8.1
idna==3.3
imageio==2.21.2
imbalanced-learn==0.9.1
importlib-metadata==4.13.0
importlib-resources==6.1.0
inequality==1.0.0
interface-meta==1.3.0
interpret==0.2.7
interpret-core==0.2.7
ipykernel==6.15.2
ipython==8.4.0
ipywidgets==8.0.1
itsdangerous==2.1.2
jarowinkler==1.2.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
JPype1==1.4.1
jsonlines==1.2.0
jsonschema==4.14.0
jupyter-dash==0.4.2
jupyter_client==7.3.5
jupyter_core==5.1.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kaleido==0.2.1
kiwisolver==1.4.4
lazy_loader==0.1rc2
libpysal==4.6.2
locket==1.0.0
lxml==4.8.0
Mako==1.2.3
mapclassify==2.4.3
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mercantile==1.2.1
Metaphone==0.6
meteostat==1.6.5
mgwr==2.1.2
mistune==2.0.4
mlflow==1.29.0
momepy==0.5.3
mplleaflet==0.0.5
mpmath==1.2.1
msgpack==1.0.5
multiprocess==0.70.13
munch==2.5.0
mypy-extensions==0.4.3
nb==0.1.2
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.4.0
ndindex==1.7
nest-asyncio==1.5.5
networkx==2.8.6
nltk==3.7
numexpr==2.8.7
numpy==1.22.4
oauthlib==3.2.1
opencv-contrib-python==4.6.0.66
openpyxl==3.0.10
openweathermap-requests==0.0.5
orjson==3.8.0
osmnet==0.1.7
osmnx==1.7.0
outdated==0.2.1
packaging==21.3
pandana==0.7
pandas==2.1.1
pandas-flavor==0.3.0
pandocfilters==1.5.0
panel==0.14.0
param==1.12.2
parso==0.8.3
partd==1.4.0
pathos==0.2.9
patsy==0.5.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pingouin==0.5.2
platformdirs==2.6.2
plotly==5.10.0
pmdarima==1.8.5
pointpats==2.2.0
pooch==1.7.0
pox==0.3.1
ppft==1.7.6.5
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.30
property-cached==1.6.4
protobuf==4.21.7
psutil==5.9.1
ptyprocess==0.7.0
PuLP==2.6.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==12.0.0
pyaurn==0.1.1
pycaret==3.0.0rc4
pycparser==2.21
pyct==0.4.8
pygeos==0.14
Pygments==2.13.0
pyhdfe==0.1.0
PyJWT==2.5.0
PyKrige==1.7.0
pynndescent==0.5.7
pyod==1.0.5
pyparsing==3.0.9
pyrobuf==0.9.3
pyrosm==0.6.1
pyrsistent==0.18.1
pysal==2.7.0
pyshp==2.3.1
python-dateutil==2.8.2
python-Levenshtein==0.12.2
python-rapidjson==1.8
pytz==2022.2.1
pyviz-comms==2.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
quantecon==0.5.3
querystring-parser==1.2.4
quilt3==5.0.0
r5py==0.1.1
racplusplus==0.1.1
rapidfuzz==2.6.1
rasterio==1.3.2
rasterstats==0.17.0
rastertodataframe==0.2.1
ratelim==0.1.6
regex==2022.8.17
requests==2.28.1
requests-cache==0.9.7
requests-futures==1.0.0
retrying==1.3.3
rich==12.6.0
rioxarray==0.12.4
Rtree==1.0.0
runipy==0.1.5
rvlib==0.0.6
s3transfer==0.6.0
SALib==1.4.5
schemdraw==0.15
scikit-gstat==1.0.1
scikit-plot==0.3.7
seaborn==0.12.2
segregation==2.3.1
setuptools-scm==6.4.2
shap==0.41.0
shapely==2.0.2
simdkalman==1.0.2
simplejson==3.17.6
six==1.16.0
skope-rules==1.0.1
slicer==0.0.7
smart-open==6.1.0
smmap==5.0.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spaghetti==1.6.6
spatialpandas==0.4.8
spglm==1.0.8
spint==1.0.7
splot==1.1.5.post1
spopt==0.4.1
spreg==1.2.4
spvcm==0.3.0
SQLAlchemy==1.4.41
sqlparse==0.4.3
stack-data==0.5.0
sympy==1.11.1
tabulate==0.8.10
tbats==1.1.0
tblib==1.7.0
tenacity==8.0.1
threadpoolctl==3.1.0
tifffile==2022.8.12
tinycss2==1.2.1
tobler==0.9.0
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.0
trace-updater==0.0.8
traitlets==5.3.0
treeinterpreter==0.2.3
trimesh==3.15.2
tsmoothie==1.0.4
typing_extensions==4.4.0
tzdata==2023.3
umap-learn==0.5.3
Unidecode==1.3.4
url-normalize==1.4.3
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.2.2
widgetsnbextension==4.0.2
wrapt==1.14.1
xarray==2022.6.0
xyzservices==2022.6.0
yellowbrick==1.5
zict==3.0.0
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
